fix: extract DropdownMenuDownloadImage from PlaygroundAction for reusability purposes
